fix(E2EReader): makes sure raw intensities are stored as float32 and not float64 to decrease the size of stored .eye files
In E2E files intensities are stored in a custom ufloat16 format this takes even less space. We actually read the data as uint16 and than map to the respective float which is converted to float32 to preserve precision. An option in future might be to store the intensities as uint16 as in E2E to further reduce file size.